+++ src/main/java/io/github/astrapi69/swing/menu/enumeration/MenuType.java
@@ -1,0 +1,72 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2021 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.swing.menu.enumeration;
+
+/**
+ * The enum class {@link MenuType} provides all the possible types that a menu can have
+ */
+public enum MenuType
+{
+
+	/**
+	 * Represents the system tray type
+	 */
+	SYSTEM_TRAY,
+
+	/**
+	 * Represents the menu bar type
+	 */
+	MENU_BAR,
+
+	/**
+	 * Represents the toolbar type
+	 */
+	TOOL_BAR,
+
+	/**
+	 * Represents the default menu type
+	 */
+	MENU,
+
+	/**
+	 * Represents the menu item type
+	 */
+	MENU_ITEM,
+
+	/**
+	 * Represents the checkbox menu item type
+	 */
+	CHECK_BOX_MENU_ITEM,
+
+	/**
+	 * Represents the radio button menu item type
+	 */
+	RADIO_BUTTON_MENU_ITEM,
+
+	/**
+	 * Represents the popup menu type
+	 */
+	POPUP
+}